Add files via upload
put the o in my name... wasnt happy with albert. 
changed  the picture of me to a picture of some html code because i seriously wasnt happy with the quality of the photo of myself i had up... this will be updated again when i find the direction id like to go, left the paragraph as a float. and i need to resize the image with % so it doesnt scroll off the side of the page. 
also updated address info, to area. so just a general idea of my area opposed to my address, we dont know these people. 
also added the new email. torresalbert@albertotorresjr.com

started working with php to get the webform to forward the message to my email.  

be back soon.